API fixes
