--- bankid-idp/bankid-idp-backend/src/main/java/se/swedenconnect/bankid/idp/NoSuchRelyingPartyException.java
@@ -1,7 +1,0 @@
-package se.swedenconnect.bankid.idp;
-
-public class NoSuchRelyingPartyException extends RuntimeException {
-  public NoSuchRelyingPartyException() {
-    super("Not registered");
-  }
-}
