--- bankid-idp/bankid-idp-backend/src/main/java/se/swedenconnect/bankid/idp/authn/ApiResponse.java
@@ -1,26 +1,0 @@
-package se.swedenconnect.bankid.idp.authn;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class ApiResponse {
-    public enum Status {
-        IN_PROGRESS,
-        ERROR,
-        COMPLETE,
-
-        CANCEL
-    }
-
-    private Status status;
-
-    private String qrCode;
-
-    private String autoStartToken;
-
-    private String messageCode;
-}
