--- bankid-idp/bankid-idp-backend/src/main/java/se/swedenconnect/bankid/idp/authn/AuthResponse.java
@@ -1,14 +1,0 @@
-package se.swedenconnect.bankid.idp.authn;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class AuthResponse {
-  private String autoStartToken;
-
-  private String qrCode;
-}
