+++ bankid-idp/bankid-idp-backend/src/main/java/se/swedenconnect/bankid/idp/authn/BankIdAuthenticationExceptionFactory.java
@@ -1,8 +1,8 @@
 package se.swedenconnect.bankid.idp.authn;
 
-import lombok.extern.slf4j.Slf4j;
-
 import java.util.UUID;
+
+import lombok.extern.slf4j.Slf4j;
 /**
  *
  * @author Martin Lindström
