+++ bankid-idp/bankid-idp-backend/src/main/java/se/swedenconnect/bankid/idp/authn/api/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * BankID IdP backend API.
+ */
+package se.swedenconnect.bankid.idp.authn.api;
