+++ bankid-idp/bankid-idp-backend/src/main/java/se/swedenconnect/bankid/idp/authn/session/BankIdSessionState.java
@@ -1,7 +1,5 @@
 package se.swedenconnect.bankid.idp.authn.session;
 
-import org.springframework.data.redis.core.TimeToLive;
-
 import java.time.Instant;
 import java.util.ArrayDeque;
 import java.util.Deque;
