+++ bankid-idp/bankid-idp-backend/src/test/resources/version.properties
@@ -1,0 +1,1 @@
+bankid.idp.version=${project.version}
